requirements.txt
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.3.0
discord-webhook==1.3.0
python-dotenv==1.0.0
